nas_check_api/requirements.txt
opencv-python
pytesseract
# sudo apt install tesseract-ocr tesseract-ocr-chi-sim
python-dotenv
